+++ src/Kiota.Builder/KiotaSearcher.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using Kiota.Builder.Configuration;
 using Kiota.Builder.SearchProviders;
 using Kiota.Builder.SearchProviders.APIsGuru;
+using Kiota.Builder.SearchProviders.GitHub;
 using Kiota.Builder.SearchProviders.MSGraph;
 using Microsoft.Extensions.Logging;
 
@@ -30,12 +31,12 @@ public class KiotaSearcher {
         var apiGurusSearchProvider = new APIsGuruSearchProvider(config.APIsGuruListUrl, client, logger, config.ClearCache);
         logger.LogDebug("searching for {searchTerm}", config.SearchTerm);
         logger.LogDebug("searching APIs.guru with url {url}", config.APIsGuruListUrl);
-        var msGraphProvider = new MSGraphSearchProvider();
         var oasProvider = new OpenApiSpecSearchProvider();
+        var githubProvider = new GitHubSearchProvider(client, config.GitHubBlockListUrl, logger, config.ClearCache);
         var results = await Task.WhenAll(
                         SearchProviderAsync(apiGurusSearchProvider, cancellationToken),
-                        SearchProviderAsync(msGraphProvider, cancellationToken),
-                        SearchProviderAsync(oasProvider, cancellationToken));
+                        SearchProviderAsync(oasProvider, cancellationToken),
+                        SearchProviderAsync(githubProvider, cancellationToken));
         return results.SelectMany(static x => x)
                 .ToDictionary(static x => x.Key, static x => x.Value, StringComparer.OrdinalIgnoreCase);
     }