+++ CHANGELOG.md
@@ -8,13 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+- Added support for GitHub based API search.[#1866](https://github.com/microsoft/kiota/issues/1866)
 - Added support for scalar request bodies Python [#1571](https://github.com/microsoft/kiota/issues/1571)
 - Sets property defaults in constructor and removes duplicate properties defined in base types from model serialization and deserialization methods in Python. [#1726](https://github.com/microsoft/kiota/issues/1726)
 - Added support for scalar request bodies in PHP [#1937](https://github.com/microsoft/kiota/pull/1937)
 - Added accept header for all schematized requests Python. [#1617](https://github.com/microsoft/kiota/issues/1617)
+- Added optional backing store support for PHP. [#1976](https://github.com/microsoft/kiota/pull/1976)
+- Fixed a bug where OdataErrors had wrong inherited class name in Python.
+- Fixed a bug where empty path parameters dictionary would throw an error in request builders in Python.
 
 ### Changed
 
+- Fixed a bug where readonly properties would fail to deserialize for TypeScript.
 - Fixed a bug where array buffers nullability would wrongly be defined for TypeScript.
 - Fixed a bug where parameter comments would appear in summary tag comments in dotnet. [#1945](https://github.com/microsoft/kiota/issues/1945)
 - Fixed a bug in PHP generation where request bodies would not serialize single elements properly. [#1937](https://github.com/microsoft/kiota/pull/1937)